refactored opcMonitor to become opcClient
