+++ edge_apps/tempusEdge_opcClient/Dockerfile
@@ -1,0 +1,5 @@
+FROM java:8-jre-alpine
+
+COPY target/opcClient-uber.jar /
+
+CMD ["java", "-jar", "/opcClient-uber.jar"]
