--- edge_apps/tempusEdge_opcMonitor/Dockerfile
@@ -1,5 +1,0 @@
-FROM java:8-jre-alpine
-
-COPY target/opcMonitor-uber.jar /
-
-CMD ["java", "-jar", "/opcMonitor-uber.jar"]
